fix: remove addSlippageToMemo in favor of addLimitToMemo
