+++ src/lib/swapper/swappers/ThorchainSwapper/utils/getL1quote.ts
@@ -25,6 +25,7 @@ import { assertUnreachable, isFulfilled, isRejected } from 'lib/utils'
 import { assertGetCosmosSdkChainAdapter } from 'lib/utils/cosmosSdk'
 import { assertGetEvmChainAdapter } from 'lib/utils/evm'
 import { THOR_PRECISION } from 'lib/utils/thorchain/constants'
+import { addLimitToMemo } from 'lib/utils/thorchain/memo/addLimitToMemo'
 import { assertGetUtxoChainAdapter } from 'lib/utils/utxo'
 import { convertDecimalPercentageToBasisPoints } from 'state/slices/tradeQuoteSlice/utils'
 
@@ -39,8 +40,8 @@ import type {
   ThorTradeUtxoOrCosmosQuote,
 } from '../getThorTradeQuote/getTradeQuote'
 import type { ThornodeQuoteResponseSuccess } from '../types'
-import { addSlippageToMemo } from './addSlippageToMemo'
 import { THORCHAIN_FIXED_PRECISION } from './constants'
+import { getLimitWithManualSlippage } from './getLimitWithManualSlippage'
 import { getQuote } from './getQuote/getQuote'
 import { TradeType } from './longTailHelpers'
 import { getEvmTxFees } from './txFeeHelpers/evmTxFees/getEvmTxFees'
@@ -221,14 +222,24 @@ export const getL1quote = async (
             const buyAmountBeforeFeesCryptoBaseUnit =
               getRouteBuyAmountBeforeFeesCryptoBaseUnit(quote)
 
-            const updatedMemo = addSlippageToMemo({
+            const limitWithManualSlippage = getLimitWithManualSlippage({
               expectedAmountOutThorBaseUnit,
-              quotedMemo: quote.memo,
               slippageBps,
-              chainId: sellAsset.chainId,
-              affiliateBps,
-              isStreaming,
             })
+
+            if (!quote.memo) throw new Error('no memo provided')
+
+            let updatedMemo = quote.memo
+
+            // always use TC auto stream quote (0 limit = 5bps - 50bps, sometimes up to 100bps)
+            // see: https://discord.com/channels/838986635756044328/1166265575941619742/1166500062101250100
+            if (!isStreaming) {
+              updatedMemo = addLimitToMemo({
+                memo: quote.memo,
+                limit: limitWithManualSlippage,
+              })
+            }
+
             const { data, router, vault } = await getEvmThorTxInfo({
               sellAsset,
               sellAmountCryptoBaseUnit,
@@ -314,14 +325,24 @@ export const getL1quote = async (
             const buyAmountBeforeFeesCryptoBaseUnit =
               getRouteBuyAmountBeforeFeesCryptoBaseUnit(quote)
 
-            const updatedMemo = addSlippageToMemo({
+            const limitWithManualSlippage = getLimitWithManualSlippage({
               expectedAmountOutThorBaseUnit,
-              quotedMemo: quote.memo,
               slippageBps,
-              isStreaming,
-              chainId: sellAsset.chainId,
-              affiliateBps,
             })
+
+            if (!quote.memo) throw new Error('no memo provided')
+
+            let updatedMemo = quote.memo
+
+            // always use TC auto stream quote (0 limit = 5bps - 50bps, sometimes up to 100bps)
+            // see: https://discord.com/channels/838986635756044328/1166265575941619742/1166500062101250100
+            if (!isStreaming) {
+              updatedMemo = addLimitToMemo({
+                memo: quote.memo,
+                limit: limitWithManualSlippage,
+              })
+            }
+
             const { vault, opReturnData, pubkey } = await getUtxoThorTxInfo({
               sellAsset,
               xpub: (input as GetUtxoTradeQuoteInput).xpub,
@@ -419,14 +440,23 @@ export const getL1quote = async (
               outputExponent: buyAsset.precision,
             }).toFixed()
 
-            const updatedMemo = addSlippageToMemo({
+            const limitWithManualSlippage = getLimitWithManualSlippage({
               expectedAmountOutThorBaseUnit,
-              quotedMemo: quote.memo,
               slippageBps,
-              isStreaming,
-              chainId: sellAsset.chainId,
-              affiliateBps,
             })
+
+            if (!quote.memo) throw new Error('no memo provided')
+
+            let updatedMemo = quote.memo
+
+            // always use TC auto stream quote (0 limit = 5bps - 50bps, sometimes up to 100bps)
+            // see: https://discord.com/channels/838986635756044328/1166265575941619742/1166500062101250100
+            if (!isStreaming) {
+              updatedMemo = addLimitToMemo({
+                memo: quote.memo,
+                limit: limitWithManualSlippage,
+              })
+            }
 
             return {
               id: uuid(),
